Updating to volunteer portal design
These pages used to be housed in iframes. However, due to 3rd party cookie blocking by some browsers, that’s not a great solution anymore. We’re now switching to just redirecting the user to a portal/community.

I’ve also updated most of the classes to use a mapper and service class.

Components used on pages have been updated to use the new layout, as well as to take advantage of inheritance to help with authentication and passing parameters.

Speaking of passing parameters, sessionStorage is now used to maintain the ID of the user moving through the portal after logging in.